pyproject.toml
[tool.poetry]
name = "tap_gmail"
version = "0.1.0"
description = ""
authors = ["Jordan <jordan.wallace.williams@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
cchardet = "^2.1.6"
python-dotenv = "^0.14.0"
pyyaml = "^5.3.1"
vcrpy = "^4.1.0"
pytest-vcr = "^1.0.2"
coverage = "^5.3"
pytest-cov = "^2.10.1"
singer-python = "^5.9.0"
requests = "^2.24.0"
google-auth = "^1.23.0"
google-auth-httplib2 = "^0.0.4"
google-api-python-client = "^1.12.5"
google-auth-oauthlib = "^0.4.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
